requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
duckdb==0.10.2
frozenlist==1.4.1
google-analytics-data==0.18.7
google-api-core==2.18.0
google-auth==2.29.0
googleapis-common-protos==1.63.0
grpcio==1.62.2
grpcio-status==1.62.2
idna==3.7
multidict==6.0.5
prettytable==3.10.0
proto-plus==1.23.0
protobuf==4.25.3
py4j==0.10.9.7
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
tabulate==0.9.0
urllib3==2.2.1
wcwidth==0.2.13
yarl==1.9.4
